fix non-root level conftest
